(updates): reduce sketch size

+++ main/src/buzzer/Buzzer.cpp
@@ -4,7 +4,7 @@ Buzzer::Buzzer(byte pin) {
     this->pin = pin;
 }
 
-void Buzzer::buzz(unsigned int frequency, unsigned long duration) {
+void Buzzer::buzz(unsigned int frequency, unsigned int duration) {
     tone(this->pin, frequency, duration);
 }
